refactor: split dbRequests into seperate database models
