+++ src/models/dbDelete.ts
@@ -1,0 +1,7 @@
+import { supabase } from "../database/supabaseClient";
+
+export async function dbDelete(table: string, id: string | number) {
+    const { data: deletedData, error } = await supabase.from(table).delete().eq('id', id);
+    if (error) throw error;
+    return deletedData || [];
+}
